moved evolution updates to core
